Make thoughts user-specific
Put paste logic in helperFunctions.js

Remove the DataSource component and the ability
for the fonrt-end to request a data source change.
