+++ src/AccountMenu.jsx
@@ -2,12 +2,13 @@ import "./CSS/Menu.css";
 import { useGlobalState } from "./GlobalStateProvider";
 
 export function AccountMenu(props) {
-  var { setUser, filters, setFilters } = useGlobalState();
+  var { setUser, loggedIn, filters, setFilters } = useGlobalState();
   return (
     <div class="menu">
       <a href="/account">Account</a>
-      <a href="/data_source">Data Source</a>
-      <a href="/thought_stack">Thought Stack</a>
+      <Show when={loggedIn()}>
+        <a href="/thought_stack">Thought Stack</a>
+      </Show>
       <a
         onClick={() => {
           setUser({});
